+++ src/features/cards/organisms/card-item.js
@@ -5,7 +5,7 @@ import { format } from "date-fns"
 
 import { RichEditor } from "@lib/rich-text"
 import { Col, Row } from "@lib/styled-components-layout"
-import { Card, H3, Link, Button, ButtonPrimary } from "@ui/atoms"
+import { Card, H3, Link, Button, ButtonPrimary } from "@ui"
 
 export const CardItem = ({ onUsefulClick, card }) => (
   <CardBox>
