+++ src/ui/index.js
@@ -1,0 +1,33 @@
+//= ======TEMPLATES=======
+export { CenterContentTemplate } from "./templates/center-content"
+export { MainTemplate } from "./templates/main-template"
+export { SidebarTemplate } from "./templates/sidebar-template"
+export { Container } from "./templates/container" // move to atoms, but check imports
+
+//= ======ORGANISMS=======
+export { PrimitiveFooter } from "./organisms/primitive-footer"
+export { ItemsList } from "./organisms/items-list"
+export { ConditionalList } from "./organisms/conditional-list"
+
+//= ======MOLECTULES=======
+export { Sidebar } from "./molecules/sidebar"
+export { TextArea } from "./molecules/textarea"
+
+//= ======ATOMS=======
+// Typography
+export { H1, H2, H3 } from "./atoms/heading"
+export { Text } from "./atoms/text"
+
+// Containers & Boxes
+export { Card, CardSticky } from "./atoms/card"
+export { ErrorBox } from "./atoms/error-box"
+
+// UI (Btns, links)
+export { Button, ButtonPrimary } from "./atoms/button"
+export { Link } from "./atoms/link"
+
+// Form Elements
+export { Input } from "./atoms/input"
+
+// Why it's here?
+export { FooterContent } from "./atoms/footer-content"
